+++ packages/cli/__tests__/zosfiles/__system__/upload/ftu/cli.files.upload.ftu.system.test.ts
@@ -74,7 +74,7 @@ describe("Upload uss file", () => {
                     defaultSys.zosmf.host,
                     defaultSys.zosmf.port,
                     defaultSys.zosmf.user,
-                    defaultSys.zosmf.password
+                    defaultSys.zosmf.password,
                 ]);
             expect(response.stderr.toString()).toBe("");
             expect(response.status).toBe(0);