+++ packages/cli/__tests__/zosjobs/__system__/delete/cli.zos-jobs.delete.job.system.test.ts
@@ -82,7 +82,7 @@ describe("zos-jobs delete job command", () => {
                         DEFAULT_SYSTEM_PROPS.zosmf.host,
                         DEFAULT_SYSTEM_PROPS.zosmf.port,
                         DEFAULT_SYSTEM_PROPS.zosmf.user,
-                        DEFAULT_SYSTEM_PROPS.zosmf.password
+                        DEFAULT_SYSTEM_PROPS.zosmf.password,
                     ]);
                 expect(response.stderr.toString()).toBe("");
                 expect(response.status).toBe(0);