+++ packages/cli/__tests__/zostso/__system__/issue/cli.zos-tso.issue.cmd.system.test.ts
@@ -103,7 +103,7 @@ describe("zos-tso issue command", () => {
                     SYSTEM_PROPS.zosmf.port,
                     SYSTEM_PROPS.zosmf.user,
                     SYSTEM_PROPS.zosmf.password,
-                    SYSTEM_PROPS.tso.account
+                    SYSTEM_PROPS.tso.account,
                 ]
             );
             expect(response.stderr.toString()).toBe("");