+++ packages/zosuss/CHANGELOG.md
@@ -22,7 +22,7 @@ All notable changes to the Zowe z/OS USS SDK package will be documented in this 
 
 ## `6.33.1`
 
-- Migrated from TSLint (now deprecated) to ESLint for static code analysis.
+- Development: Migrated from TSLint (now deprecated) to ESLint for static code analysis.
 
 ## `6.32.1`