+++ packages/zosuss/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe z/OS USS SDK package will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Ensure `zowe config auto-init` command saves the `rejectUnauthorized` value. [#1109](https://github.com/zowe/zowe-cli/issues/1109)
+
 ## `7.0.0-next.202106071827`
 
 - **Breaking**: Removed the following [deprecated API functions](https://github.com/zowe/zowe-cli/pull/1022):