+++ packages/zosuss/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to the Zowe z/OS USS SDK package will be documented in this file.
 
+## Recent Changes
+
+- Documentation: Deprecated the following command option definitions:
+  - SSH_OPTION_HOST_PROFILE
+    - Replace with SSH_OPTION_HOST
+  - SSH_OPTION_USER_PROFILE
+    - Replace with SSH_OPTION_USER
+
 ## `7.0.0-next.202106071827`
 
 - **Breaking**: Removed the following [deprecated API functions](https://github.com/zowe/zowe-cli/pull/1022):